[AAP-71795] Move EmailAdminOnlyMixin tests to dedicated file
Tests the mixin in isolation using a minimal EmailTestSerializer
rather than through CommonUserSerializer/UserSerializer.
